+++ dcsbridge/__main__.py
@@ -3,11 +3,12 @@ from dcsbridge.dataloaders.combatflite import MissionPlanDataLoader
 from dcsbridge.dataloaders.common import DataLoader
 from dcsbridge.dataloaders.scratchpad import ScratchpadDataLoader
 from time import localtime
-from pathlib import Path
 import sys
 import socket
+import argparse
 
-__default_bingo = "4000"
+__DEFAULT_BINGO = "4000"
+__PROGRAM_NAME = "dcs-bridge"
 
 
 class DcsBios:
@@ -20,50 +21,77 @@ class DcsBios:
         self.__connection.sendto(msg.encode("utf-8"), (self.__host, self.__port))
 
 
+def execute_mission(driver, args):
+    dl = MissionPlanDataLoader()
+    dl.load_data()
+    driver.enter_steerpoints(dl.get_waypoints())
+    driver.enter_bingo(dl.get_bingo())
+
+
+def execute_bingo(driver, args):
+    driver.enter_bingo(args.bingo)
+
+
+def execute_time(driver, args):
+    if "local" == args.timezone:
+        driver.enter_time(localtime())
+
+
+def execute_index(driver, args):
+    dl = DataLoader(args.file)
+    dl.load_data()
+    driver.enter_steerpoint(dl.get_waypoint(args.index))
+
+
+def execute_scratchpad(driver, args):
+    dl = ScratchpadDataLoader(args.file)
+    dl.load_data()
+    driver.enter_steerpoints(dl.get_waypoints())
+
+
+def parse_arguments(argv):
+    parser = argparse.ArgumentParser(prog=__PROGRAM_NAME)
+    subparser = parser.add_subparsers(dest="command")
+    mission = subparser.add_parser("mission")
+    bingo = subparser.add_parser("bingo")
+    time = subparser.add_parser("time")
+    index = subparser.add_parser("index")
+    scratchpad = subparser.add_parser("scratchpad")
+
+    mission.add_argument("file", type=str, help="path to excel file with mission plan")
+    mission.add_argument("--default-bingo", type=int, required=False, default=__DEFAULT_BINGO,
+                         help="default bingo value if mission from mission plan")
+
+    bingo.add_argument("bingo", type=int, help="bingo value to set in airframe")
+
+    time.add_argument("timezone", type=str, default='local', choices=['local'],
+                      help="timezone to use when entering current time")
+
+    index.add_argument("file", type=str, help="path to data file")
+    index.add_argument("index", type=int, help="index value of entry to load")
+
+    scratchpad.add_argument("--file", type=str, required=False,
+        help="file where DCS Scratchpad coordinates are stored, defaults to 0000.txt in default Scratchpad")
+
+    return parser.parse_args(argv)
+
+
 def main(argv):
-    num_args = len(argv)
-    if num_args < 2:
-        raise Exception("No commandline argument")
+    args = parse_arguments(argv)
 
     dcsbios = DcsBios()
     d = Driver(dcsbios)
 
-    if "mission" == argv[1]:  # Load mission plan from CombatFlite: mission
-        dl = MissionPlanDataLoader()
-        dl.load_data()
-        d.enter_steerpoints(dl.get_waypoints())
-        d.enter_bingo(dl.get_bingo())
-    elif "bingo" == argv[1]:  # Set bingo value: bingo <value>
-        if num_args == 3:
-            d.enter_bingo(argv[2])
-        else:
-            d.enter_bingo(__default_bingo)
-    elif "time" == argv[1]:  # Set aircraft system time to local time: time
-        d.enter_time(localtime())
-    elif "index" == argv[1]:  # Load objective index: index <data file> <index>
-        if num_args < 4:
-            raise Exception("Invalid arguments: index <data file> <index>")
+    handlers = {
+        "mission": execute_mission,
+        "bingo": execute_bingo,
+        "time": execute_time,
+        "index": execute_index,
+        "scratchpad": execute_scratchpad
+    }
 
-        data_file = argv[2]
-        index = argv[3]
-
-        dl = DataLoader(data_file)
-        dl.load_data()
-        d.enter_steerpoint(dl.get_waypoint(index))
-    elif "scratchpad" == argv[1]: # Load steerpoints from DCS Scratchpad files
-        data_file = Path("Scratchpad") / Path("0000.txt")
-        if num_args == 3:
-            data_file = Path(argv[2])
-
-        dl = ScratchpadDataLoader(data_file)
-        dl.load_data()
-        d.enter_steerpoints(dl.get_waypoints())
-    else:
-        raise Exception("Unsupported operation")
+    handlers.get(args.command)(d, args)
 
 
-try:
-    main(sys.argv)
-except Exception:
-    raise
-
+if __name__ == "__main__":
+    main(sys.argv[1:])